--- src/main/kotlin/com/capstone/jachulsa/config/CorsConfig.kt
@@ -1,23 +1,0 @@
-package com.capstone.jachulsa.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.cors.CorsConfiguration
-import org.springframework.web.cors.CorsConfigurationSource
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource
-
-@Configuration
-class CorsConfig{
-    @Bean
-    fun corsConfigurationSource(): CorsConfigurationSource {
-        val configuration = CorsConfiguration()
-        configuration.allowCredentials = true;
-        configuration.allowedOrigins = listOf("http://localhost:3000","https://jachulsa.shop","https://dev.jachulsa.shop");
-        configuration.allowedMethods = listOf("GET", "POST", "PUT", "DELETE", "PATCH", "OPTIONS");
-        configuration.allowedHeaders = listOf("*");
-        configuration.exposedHeaders = listOf("*");
-        val source = UrlBasedCorsConfigurationSource()
-        source.registerCorsConfiguration("/**", configuration)
-        return source
-    }
-}
